refactor: isolate NVD path resolution
Moves durable config, state/local-data, and taxonomy path resolution into a state-neutral py_nvd.paths module. The old py_nvd.db names remain as compatibility re-exports so existing callers keep working while the state database module is decomposed.

Updates taxonomy and CLI config lookup code to import path helpers from the new boundary. This preserves NVD_CONFIG, NVD_STATE_DIR, and NVD_TAXONOMY_DB behavior before removing the old state and hit surfaces.
